test: move stub components in to org.terasology.unittest module.
