--- engine-tests/src/main/java/org/terasology/engine/entitySystem/stubs/ListOfEnumsComponent.java
@@ -1,26 +1,0 @@
-/*
- * Copyright 2017 MovingBlocks
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.terasology.engine.entitySystem.stubs;
-
-import org.terasology.engine.entitySystem.Component;
-import org.terasology.engine.math.Side;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class ListOfEnumsComponent implements Component {
-    public List<Side> elements = new ArrayList<>();
-}
