--- engine-tests/src/main/java/org/terasology/engine/entitySystem/stubs/MappedContainerComponent.java
@@ -1,30 +1,0 @@
-/*
- * Copyright 2015 MovingBlocks
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.terasology.engine.entitySystem.stubs;
-
-import org.terasology.engine.entitySystem.Component;
-import org.terasology.reflection.MappedContainer;
-
-import java.util.Set;
-
-public class MappedContainerComponent implements Component {
-    public Set<Cont> containers;
-
-    @MappedContainer
-    public static class Cont {
-        public String value;
-    }
-}
