+++ engine-tests/src/main/java/org/terasology/unittest/stubs/ForceBlockActiveComponent.java
@@ -1,0 +1,12 @@
+// Copyright 2021 The Terasology Foundation
+// SPDX-License-Identifier: Apache-2.0
+package org.terasology.unittest.stubs;
+
+import org.terasology.engine.entitySystem.Component;
+import org.terasology.engine.world.block.ForceBlockActive;
+
+/**
+ */
+@ForceBlockActive
+public class ForceBlockActiveComponent implements Component {
+}
