+++ engine-tests/src/main/java/org/terasology/unittest/stubs/OwnerComponent.java
@@ -1,0 +1,14 @@
+// Copyright 2021 The Terasology Foundation
+// SPDX-License-Identifier: Apache-2.0
+package org.terasology.unittest.stubs;
+
+import org.terasology.engine.entitySystem.Component;
+import org.terasology.engine.entitySystem.Owns;
+import org.terasology.engine.entitySystem.entity.EntityRef;
+
+/**
+ */
+public class OwnerComponent implements Component {
+    @Owns
+    public EntityRef child = EntityRef.NULL;
+}
