+++ engine-tests/src/main/java/org/terasology/unittest/stubs/StringComponent.java
@@ -1,6 +1,6 @@
 // Copyright 2021 The Terasology Foundation
 // SPDX-License-Identifier: Apache-2.0
-package org.terasology.engine.entitySystem.stubs;
+package org.terasology.unittest.stubs;
 
 import com.google.common.base.MoreObjects;
 import com.google.common.base.Objects;
