+++ engine-tests/src/test/java/org/terasology/engine/entitySystem/PrefabTest.java
@@ -19,11 +19,11 @@ import org.terasology.engine.entitySystem.prefab.PrefabManager;
 import org.terasology.engine.entitySystem.prefab.internal.PojoPrefab;
 import org.terasology.engine.entitySystem.prefab.internal.PojoPrefabManager;
 import org.terasology.engine.entitySystem.prefab.internal.PrefabFormat;
-import org.terasology.engine.entitySystem.stubs.ListOfEnumsComponent;
-import org.terasology.engine.entitySystem.stubs.ListOfObjectComponent;
-import org.terasology.engine.entitySystem.stubs.MappedContainerComponent;
-import org.terasology.engine.entitySystem.stubs.OrderedMapTestComponent;
-import org.terasology.engine.entitySystem.stubs.StringComponent;
+import org.terasology.unittest.stubs.ListOfEnumsComponent;
+import org.terasology.unittest.stubs.ListOfObjectComponent;
+import org.terasology.unittest.stubs.MappedContainerComponent;
+import org.terasology.unittest.stubs.OrderedMapTestComponent;
+import org.terasology.unittest.stubs.StringComponent;
 import org.terasology.engine.math.Side;
 import org.terasology.engine.network.NetworkMode;
 import org.terasology.engine.network.NetworkSystem;
