+++ engine-tests/src/test/java/org/terasology/engine/entitySystem/metadata/ComponentMetadataTest.java
@@ -1,4 +1,4 @@
-// Copyright 2020 The Terasology Foundation
+// Copyright 2021 The Terasology Foundation
 // SPDX-License-Identifier: Apache-2.0
 
 package org.terasology.engine.entitySystem.metadata;
@@ -9,8 +9,8 @@ import org.reflections.Reflections;
 import org.terasology.gestalt.assets.ResourceUrn;
 import org.terasology.engine.context.Context;
 import org.terasology.engine.context.internal.ContextImpl;
-import org.terasology.engine.entitySystem.stubs.OwnerComponent;
-import org.terasology.engine.entitySystem.stubs.StringComponent;
+import org.terasology.unittest.stubs.OwnerComponent;
+import org.terasology.unittest.stubs.StringComponent;
 import org.terasology.engine.persistence.typeHandling.TypeHandlerLibraryImpl;
 import org.terasology.reflection.copy.CopyStrategyLibrary;
 import org.terasology.reflection.reflect.ReflectFactory;
