+++ engine-tests/src/test/java/org/terasology/engine/persistence/ComponentSerializerTest.java
@@ -17,9 +17,9 @@ import org.terasology.engine.core.bootstrap.EntitySystemSetupUtil;
 import org.terasology.engine.entitySystem.Component;
 import org.terasology.engine.entitySystem.entity.internal.EngineEntityManager;
 import org.terasology.engine.entitySystem.metadata.ComponentLibrary;
-import org.terasology.engine.entitySystem.stubs.GetterSetterComponent;
-import org.terasology.engine.entitySystem.stubs.IntegerComponent;
-import org.terasology.engine.entitySystem.stubs.StringComponent;
+import org.terasology.unittest.stubs.GetterSetterComponent;
+import org.terasology.unittest.stubs.IntegerComponent;
+import org.terasology.unittest.stubs.StringComponent;
 import org.terasology.engine.network.NetworkMode;
 import org.terasology.engine.network.NetworkSystem;
 import org.terasology.engine.persistence.serializers.ComponentSerializer;
