+++ engine-tests/src/test/java/org/terasology/engine/persistence/internal/StorageManagerTest.java
@@ -23,8 +23,8 @@ import org.terasology.engine.core.module.ModuleManager;
 import org.terasology.engine.core.paths.PathManager;
 import org.terasology.engine.entitySystem.entity.EntityRef;
 import org.terasology.engine.entitySystem.entity.internal.EngineEntityManager;
-import org.terasology.engine.entitySystem.stubs.EntityRefComponent;
-import org.terasology.engine.entitySystem.stubs.StringComponent;
+import org.terasology.unittest.stubs.EntityRefComponent;
+import org.terasology.unittest.stubs.StringComponent;
 import org.terasology.joml.geom.AABBfc;
 import org.terasology.engine.logic.location.LocationComponent;
 import org.terasology.gestalt.module.ModuleEnvironment;
